3.0.12.1 UNSTABLE -- huge update, lots missing still
NOTE: A lot of this code is untested and I most likely broke a lot of
interactions that have been delicately added over the last year.  There
will need to be substantial testing and revision to get the game def
running on the same level it did before.

- Merged trigAbility and stackResolve into a single function
(autoParser)
- Activate Ability popup window now has Radio buttons and shows full
rules text, autoscripted abilities show up with a blue background
- New autoscript online repository, is currently empty so no cards are
autoscripted yet
- webRead functions are more advanced and safer to use