New action: Take from library top X, shuffle rest
Fixes #125.

This adds a new action, which allows the user to have a look at top X cards from their library, take any number of them to hand using a Scry-like window (with the option of revealing these cards to all opponents), and then moves the rest to the bottom of library in a random order.

Effects of this kind are recently quite popular and until now resolving them was very inconvenient on OCTGN, as explained by #125.